refactor: move shared hud components
