+++ src/components/index.ts
@@ -1,0 +1,9 @@
+export { AppButton } from './AppButton';
+export { BossHealthBar } from './BossHealthBar';
+export { EncounterBrand } from './EncounterBrand';
+export { Modal } from './Modal';
+export { SurfaceSection } from './SurfaceSection';
+
+export type { BossHealthBarProps } from './BossHealthBar';
+export type { EncounterBrandProps } from './EncounterBrand';
+export type { ModalProps } from './Modal';
